refactor: split extractor into focused modules
