--- graph_pdf/extractor.py
@@ -1,1940 +1,0 @@
-from __future__ import annotations
-
-import json
-import math
-import re
-from pathlib import Path
-from typing import List, Optional, Sequence, Tuple
-
-import pdfplumber
-from pypdf import PdfReader
-
-TableRows = List[List[str]]
-TableChunk = Tuple[TableRows, Tuple[float, float, float, float]]
-WATERMARK_ROTATION_MIN_DEGREES = 53.0
-WATERMARK_ROTATION_MAX_DEGREES = 57.0
-WATERMARK_GRAY_MIN = 0.88
-WATERMARK_GRAY_MAX = 0.96
-WATERMARK_GRAY_NEUTRAL_TOLERANCE = 0.03
-BULLET_PREFIX_RE = re.compile(
-    r"^(?:[-*•●○◦◯▪▫■□◆◇◈◊‣∙◉]|[0-9]+(?:[-.][0-9]+)*[.)]|o|\?|\uFFFD)\s+"
-)
-
-
-def _parse_pages_spec(spec: str) -> List[int]:
-    values = set()
-    for part in str(spec or "").split(","):
-        token = part.strip()
-        if not token:
-            raise ValueError("empty page token in --pages")
-        if "-" in token:
-            start_text, end_text = token.split("-", 1)
-            if not start_text.isdigit() or not end_text.isdigit():
-                raise ValueError(f"invalid page range: {token}")
-            start = int(start_text)
-            end = int(end_text)
-            if start < 1 or end < 1 or start > end:
-                raise ValueError(f"invalid page range: {token}")
-            values.update(range(start, end + 1))
-            continue
-        if not token.isdigit():
-            raise ValueError(f"invalid page number: {token}")
-        page_no = int(token)
-        if page_no < 1:
-            raise ValueError(f"invalid page number: {token}")
-        values.add(page_no)
-    if not values:
-        raise ValueError("no pages selected")
-    return sorted(values)
-
-
-def _normalize_text(text: str) -> str:
-    return re.sub(r"\s+", " ", text or "").strip()
-
-def _char_rotation_degrees(char: dict) -> float:
-    matrix = char.get("matrix")
-    if not isinstance(matrix, tuple) or len(matrix) < 2:
-        return 0.0
-    return math.degrees(math.atan2(float(matrix[1]), float(matrix[0])))
-
-
-def _collect_rotated_text_debug(page: "pdfplumber.page.Page", page_no: int) -> List[dict]:
-    entries: List[dict] = []
-    for char in getattr(page, "chars", []):
-        rotation = _char_rotation_degrees(char)
-        if abs(rotation) <= 0.1:
-            continue
-        entries.append(
-            {
-                "page": page_no,
-                "text": str(char.get("text", "")),
-                "rotation": rotation,
-                "size": float(char.get("size", 0.0)),
-                "x0": float(char.get("x0", 0.0)),
-                "x1": float(char.get("x1", 0.0)),
-                "top": float(char.get("top", 0.0)),
-                "bottom": float(char.get("bottom", 0.0)),
-                "matrix": list(char.get("matrix", ())),
-                "non_stroking_color": char.get("non_stroking_color"),
-                "stroking_color": char.get("stroking_color"),
-                "fontname": char.get("fontname"),
-            }
-        )
-    return entries
-
-
-def _merge_numeric_positions(values: Sequence[float], tolerance: float = 1.0) -> List[float]:
-    merged: List[float] = []
-    for value in sorted(float(v) for v in values):
-        if not merged or abs(value - merged[-1]) > tolerance:
-            merged.append(value)
-            continue
-        merged[-1] = (merged[-1] + value) / 2.0
-    return merged
-
-
-def _cluster_axis_values(values: Sequence[float], tolerance: float = 1.0) -> List[List[float]]:
-    clusters: List[List[float]] = []
-    for value in sorted(float(v) for v in values):
-        if not clusters or abs(value - clusters[-1][-1]) > tolerance:
-            clusters.append([value])
-            continue
-        clusters[-1].append(value)
-    return clusters
-
-
-def _round_segment(
-    edge: dict,
-    body_top: float | None = None,
-    body_bottom: float | None = None,
-) -> dict:
-    payload = {
-        "x0": round(float(edge["x0"]), 2),
-        "x1": round(float(edge["x1"]), 2),
-        "top": round(float(edge["top"]), 2),
-        "bottom": round(float(edge["bottom"]), 2),
-    }
-    if body_top is not None and body_bottom is not None:
-        payload["in_body_bounds"] = (
-            float(edge["bottom"]) > body_top and float(edge["top"]) < body_bottom
-        )
-    return payload
-
-
-def _normalize_band_segments(
-    segments: Sequence[dict],
-    start_key: str,
-    end_key: str,
-    orth_min_key: str,
-    orth_max_key: str,
-    tolerance: float = 1.0,
-) -> List[dict]:
-    normalized: List[dict] = []
-    ordered = sorted(
-        (dict(edge) for edge in segments),
-        key=lambda item: (float(item[start_key]), float(item[end_key])),
-    )
-    for edge in ordered:
-        if not normalized:
-            normalized.append(edge)
-            continue
-
-        current = normalized[-1]
-        current_start = float(current[start_key])
-        current_end = float(current[end_key])
-        edge_start = float(edge[start_key])
-        edge_end = float(edge[end_key])
-
-        # Fully-contained duplicates should only widen the orthogonal span.
-        if edge_start >= current_start and edge_end <= current_end:
-            current[orth_min_key] = min(float(current[orth_min_key]), float(edge[orth_min_key]))
-            current[orth_max_key] = max(float(current[orth_max_key]), float(edge[orth_max_key]))
-            continue
-
-        if edge_start - current_end <= tolerance:
-            current[end_key] = max(current_end, edge_end)
-            current[orth_min_key] = min(float(current[orth_min_key]), float(edge[orth_min_key]))
-            current[orth_max_key] = max(float(current[orth_max_key]), float(edge[orth_max_key]))
-            continue
-
-        normalized.append(edge)
-
-    return normalized
-
-
-def _merge_horizontal_band_segments(
-    segments: Sequence[dict],
-    tolerance: float = 1.0,
-    body_top: float | None = None,
-    body_bottom: float | None = None,
-) -> List[dict]:
-    merged = _normalize_band_segments(
-        segments,
-        start_key="x0",
-        end_key="x1",
-        orth_min_key="top",
-        orth_max_key="bottom",
-        tolerance=tolerance,
-    )
-    return [_round_segment(edge, body_top=body_top, body_bottom=body_bottom) for edge in merged]
-
-
-def _merge_vertical_band_segments(
-    segments: Sequence[dict],
-    tolerance: float = 1.0,
-    body_top: float | None = None,
-    body_bottom: float | None = None,
-) -> List[dict]:
-    merged = _normalize_band_segments(
-        segments,
-        start_key="top",
-        end_key="bottom",
-        orth_min_key="x0",
-        orth_max_key="x1",
-        tolerance=tolerance,
-    )
-    return [_round_segment(edge, body_top=body_top, body_bottom=body_bottom) for edge in merged]
-
-
-def _build_segment_groups(
-    segments: Sequence[dict],
-    axis_key: str,
-    merge_fn,
-    tolerance: float = 1.0,
-    body_top: float | None = None,
-    body_bottom: float | None = None,
-) -> List[dict]:
-    clusters = _cluster_axis_values([float(edge[axis_key]) for edge in segments], tolerance=tolerance)
-    groups: List[dict] = []
-    for cluster in clusters:
-        axis = sum(cluster) / len(cluster)
-        members = [
-            edge
-            for edge in segments
-            if any(abs(float(edge[axis_key]) - value) <= tolerance for value in cluster)
-        ]
-        groups.append(
-            {
-                "axis": round(axis, 2),
-                "segments": [
-                    _round_segment(edge, body_top=body_top, body_bottom=body_bottom)
-                    for edge in members
-                ],
-                "merged_segments": merge_fn(
-                    members,
-                    tolerance=tolerance,
-                    body_top=body_top,
-                    body_bottom=body_bottom,
-                ),
-            }
-        )
-    return groups
-
-
-def _collect_table_drawing_debug(
-    page: "pdfplumber.page.Page",
-    page_no: int,
-    header_margin: float = 90.0,
-    footer_margin: float = 40.0,
-) -> dict:
-    body_top, body_bottom = _detect_body_bounds(
-        page,
-        header_margin=header_margin,
-        footer_margin=footer_margin,
-    )
-    groups = _table_regions(page)
-    tables: List[dict] = []
-    for index, (x0, x1, lines) in enumerate(groups, start=1):
-        top = min(float(edge["top"]) for edge in lines)
-        bottom = max(float(edge["top"]) for edge in lines)
-        horizontal_edges = [
-            edge
-            for edge in lines
-            if float(edge["top"]) >= body_top
-            and float(edge["bottom"]) <= body_bottom
-        ]
-        vertical_edges = [
-            edge
-            for edge in page.vertical_edges
-            if float(edge["x0"]) >= x0 - 2.0
-            and float(edge["x0"]) <= x1 + 2.0
-            and float(edge["bottom"]) >= top
-            and float(edge["top"]) <= bottom
-        ]
-        horizontal_positions = _merge_numeric_positions([float(edge["top"]) for edge in horizontal_edges])
-        vertical_positions = _merge_numeric_positions(
-            [x0, x1, *(float(edge["x0"]) for edge in vertical_edges)]
-        )
-        horizontal_groups = _build_segment_groups(
-            horizontal_edges,
-            axis_key="top",
-            merge_fn=_merge_horizontal_band_segments,
-            tolerance=1.0,
-            body_top=body_top,
-            body_bottom=body_bottom,
-        )
-        vertical_groups = _build_segment_groups(
-            vertical_edges,
-            axis_key="x0",
-            merge_fn=_merge_vertical_band_segments,
-            tolerance=1.0,
-            body_top=body_top,
-            body_bottom=body_bottom,
-        )
-        tables.append(
-            {
-                "index": index,
-                "bbox": [round(x0, 2), round(top, 2), round(x1, 2), round(bottom, 2)],
-                "row_count": max(0, len(horizontal_positions) - 1),
-                "col_count": max(0, len(vertical_positions) - 1),
-                "horizontal_lines": [round(value, 2) for value in horizontal_positions],
-                "vertical_lines": [round(value, 2) for value in vertical_positions],
-                "horizontal_segments": [
-                    {
-                        **_round_segment(edge, body_top=body_top, body_bottom=body_bottom),
-                    }
-                    for edge in horizontal_edges
-                ],
-                "vertical_segments": [
-                    {
-                        **_round_segment(edge, body_top=body_top, body_bottom=body_bottom),
-                    }
-                    for edge in vertical_edges
-                ],
-                "horizontal_groups": horizontal_groups,
-                "vertical_groups": vertical_groups,
-                "horizontal_count": len(horizontal_positions),
-                "vertical_count": len(vertical_positions),
-            }
-        )
-
-    line_payloads = _extract_body_word_lines(
-        page=page,
-        header_margin=header_margin,
-        footer_margin=footer_margin,
-    )
-    raw_text_lines, normalized_text_lines = _extract_body_text_lines(
-        page,
-        header_margin=header_margin,
-        footer_margin=footer_margin,
-    )
-
-    return {
-        "page": page_no,
-        "body_bounds": [round(body_top, 2), round(body_bottom, 2)],
-        "table_count": len(tables),
-        "tables": tables,
-        "text_debug": {
-            "raw_lines": raw_text_lines,
-            "raw_line_boxes": [
-                {
-                    "text": str(line.get("text") or ""),
-                    "x0": round(float(line.get("x0", 0.0)), 2),
-                    "x1": round(float(line.get("x1", 0.0)), 2),
-                    "top": round(float(line.get("top", 0.0)), 2),
-                    "bottom": round(float(line.get("bottom", 0.0)), 2),
-                    "text_start_x": round(float(line.get("text_start_x", line.get("x0", 0.0))), 2),
-                    "marker_candidate": bool(line.get("marker_candidate")),
-                }
-                for line in line_payloads
-            ],
-            "normalized_lines": normalized_text_lines,
-        },
-    }
-
-
-def _collect_page_edge_debug(
-    page: "pdfplumber.page.Page",
-    page_no: int,
-    header_margin: float = 90.0,
-    footer_margin: float = 40.0,
-) -> dict:
-    body_top, body_bottom = _detect_body_bounds(
-        page,
-        header_margin=header_margin,
-        footer_margin=footer_margin,
-    )
-    groups = _table_regions(page)
-    selected_horizontal_edges = []
-    selected_vertical_edges = []
-    for x0, x1, lines in groups:
-        top = min(float(edge["top"]) for edge in lines)
-        bottom = max(float(edge["top"]) for edge in lines)
-        selected_horizontal_edges.extend(lines)
-        selected_vertical_edges.extend(
-            edge
-            for edge in page.vertical_edges
-            if float(edge["x0"]) >= x0 - 2.0
-            and float(edge["x0"]) <= x1 + 2.0
-            and float(edge["bottom"]) >= top
-            and float(edge["top"]) <= bottom
-        )
-
-    return {
-        "page": page_no,
-        "body_bounds": [round(body_top, 2), round(body_bottom, 2)],
-        "all_horizontal_edges": [
-            _round_segment(edge, body_top=body_top, body_bottom=body_bottom)
-            for edge in page.horizontal_edges
-        ],
-        "selected_horizontal_edges": [
-            _round_segment(edge, body_top=body_top, body_bottom=body_bottom)
-            for edge in selected_horizontal_edges
-        ],
-        "all_vertical_edges": [
-            _round_segment(edge, body_top=body_top, body_bottom=body_bottom)
-            for edge in page.vertical_edges
-        ],
-        "selected_vertical_edges": [
-            _round_segment(edge, body_top=body_top, body_bottom=body_bottom)
-            for edge in selected_vertical_edges
-        ],
-    }
-
-
-def _repair_watermark_bleed(text: str) -> str:
-    # Some rotated/conflicting watermark text can leak as a trailing single letter.
-    text = re.sub(r"\s+[A-Za-z]$", "", text)
-    return text.strip()
-
-
-def _is_layout_artifact(text: str) -> bool:
-    normalized = _normalize_text(text).lower()
-    if not normalized:
-        return True
-
-    if "graph pdf demo header" in normalized:
-        return True
-
-    if "chapter 1: deep structure verification" in normalized:
-        return False
-
-    if re.search(r"^page \d+\s*/\s*\d+$", normalized):
-        return True
-
-    header_markers = (
-        "prepared for table + text extraction tests",
-        "header checks:",
-        "header line",
-        "header line 1",
-        "header line 2",
-        "header line 3",
-    )
-    footer_markers = (
-        "graph pdf demo footer / left",
-        "footer details: keep header/footer clean",
-        "footer note: ignore this for body extraction",
-        "footer line 1:",
-        "footer line 2:",
-        "footer line 3:",
-        "footer line marker:",
-        "footer page marker:",
-    )
-    return any(marker in normalized for marker in (*header_markers, *footer_markers))
-
-def _is_gray_color(color: object) -> bool:
-    if not isinstance(color, tuple) or len(color) < 3:
-        return False
-    rgb = [float(c) for c in color[:3]]
-    brightness = sum(rgb) / 3.0
-    return max(rgb) - min(rgb) <= WATERMARK_GRAY_NEUTRAL_TOLERANCE and WATERMARK_GRAY_MIN <= brightness <= WATERMARK_GRAY_MAX
-
-
-def _is_non_watermark_obj(obj: dict) -> bool:
-    if obj.get("object_type") != "char":
-        return True
-
-    angle = _char_rotation_degrees(obj)
-    color = obj.get("non_stroking_color") or obj.get("stroking_color")
-    is_gray_watermark = WATERMARK_ROTATION_MIN_DEGREES <= angle <= WATERMARK_ROTATION_MAX_DEGREES and _is_gray_color(color)
-    return not is_gray_watermark
-
-
-def _filter_page_for_extraction(page: "pdfplumber.page.Page") -> "pdfplumber.page.Page":
-    return page.filter(_is_non_watermark_obj)
-
-
-def _detect_chapter_body_top(page: "pdfplumber.page.Page") -> float | None:
-    extract_words = getattr(page, "extract_words", None)
-    if not callable(extract_words):
-        return None
-
-    words = extract_words(
-        x_tolerance=1.5,
-        y_tolerance=2.0,
-        keep_blank_chars=False,
-        extra_attrs=["size", "fontname"],
-    ) or []
-    if not words:
-        return None
-
-    grouped_lines: List[List[dict]] = []
-    for word in sorted(words, key=lambda item: (float(item.get("top", 0.0)), float(item.get("x0", 0.0)))):
-        if not grouped_lines or abs(float(word.get("top", 0.0)) - float(grouped_lines[-1][0].get("top", 0.0))) > 2.5:
-            grouped_lines.append([word])
-            continue
-        grouped_lines[-1].append(word)
-
-    page_top_limit = float(page.height) * 0.25
-    for words_in_line in grouped_lines:
-        ordered = sorted(words_in_line, key=lambda item: float(item.get("x0", 0.0)))
-        text = " ".join(str(word.get("text") or "").strip() for word in ordered).strip()
-        if not text:
-            continue
-        top = min(float(word.get("top", 0.0)) for word in ordered)
-        if top > page_top_limit:
-            continue
-        avg_size = sum(float(word.get("size", 0.0)) for word in ordered) / max(len(ordered), 1)
-        if avg_size < 20.0:
-            continue
-        if not re.search(r"\b(chapter|section|appendix)\b", text, flags=re.IGNORECASE):
-            continue
-        return top
-
-    return None
-
-
-def _detect_body_bounds(
-    page: "pdfplumber.page.Page",
-    header_margin: float,
-    footer_margin: float,
-) -> Tuple[float, float]:
-    default_top = float(footer_margin)
-    default_bottom = float(page.height - header_margin)
-    min_width = float(page.width) * 0.7
-
-    top_candidates = [
-        edge
-        for edge in getattr(page, "horizontal_edges", [])
-        if float(edge.get("x1", 0.0)) - float(edge.get("x0", 0.0)) >= min_width
-        and float(edge.get("top", 0.0)) <= float(page.height) * 0.2
-    ]
-    bottom_candidates = [
-        edge
-        for edge in getattr(page, "horizontal_edges", [])
-        if float(edge.get("x1", 0.0)) - float(edge.get("x0", 0.0)) >= min_width
-        and float(edge.get("top", 0.0)) >= float(page.height) * 0.8
-    ]
-
-    body_top = min((float(edge["top"]) for edge in top_candidates), default=default_top)
-    body_bottom = max((float(edge["top"]) for edge in bottom_candidates), default=default_bottom)
-
-    if not top_candidates:
-        chapter_top = _detect_chapter_body_top(page)
-        if chapter_top is not None:
-            body_top = chapter_top
-
-    if body_bottom <= body_top:
-        return default_top, default_bottom
-    return body_top, body_bottom
-
-
-def _extract_body_text(
-    page: "pdfplumber.page.Page",
-    header_margin: float,
-    footer_margin: float,
-    excluded_bboxes: Sequence[Tuple[float, float, float, float]] = (),
-) -> str:
-    _raw_lines, normalized_lines = _extract_body_text_lines(
-        page=page,
-        header_margin=header_margin,
-        footer_margin=footer_margin,
-        excluded_bboxes=excluded_bboxes,
-    )
-    return "\n".join(normalized_lines)
-
-
-def _extract_body_text_lines(
-    page: "pdfplumber.page.Page",
-    header_margin: float,
-    footer_margin: float,
-    excluded_bboxes: Sequence[Tuple[float, float, float, float]] = (),
-) -> Tuple[List[str], List[str]]:
-    line_payloads = _extract_body_word_lines(
-        page=page,
-        header_margin=header_margin,
-        footer_margin=footer_margin,
-        excluded_bboxes=excluded_bboxes,
-    )
-    raw_lines = [str(line["text"]) for line in line_payloads]
-    blocks = _build_body_blocks(line_payloads)
-
-    normalized_lines: List[str] = []
-    for block in blocks:
-        block_lines = [str(line["text"]) for line in block["lines"]]
-        if block["kind"] == "heading":
-            normalized_lines.extend(block_lines)
-            continue
-        joined = _join_non_heading_block_lines(block_lines)
-        if joined:
-            normalized_lines.append(joined)
-
-    return raw_lines, normalized_lines
-
-
-def _merge_cells(table: Sequence[Sequence[str]]) -> List[List[str]]:
-    out = []
-    for row in table:
-        out.append([str(cell or "").strip() for cell in row])
-    return out
-
-
-def _clean_cell_line(line: str) -> str:
-    cleaned = str(line or "").strip()
-    cleaned = re.sub(r"\s+", " ", cleaned).strip()
-    tokens = cleaned.split()
-    if len(tokens) >= 2 and tokens[-1].upper() == "I":
-        cleaned = " ".join(tokens[:-1]).strip()
-    return cleaned
-
-
-def _remove_watermark_fragment_lines(lines: Sequence[str]) -> List[str]:
-    cleaned = [_clean_cell_line(line) for line in lines]
-    cleaned = [line for line in cleaned if line]
-
-    if len(cleaned) <= 1:
-        return cleaned
-
-    return cleaned
-
-
-def _is_bullet_line(line: str) -> bool:
-    return bool(BULLET_PREFIX_RE.match(str(line or "").strip()))
-
-
-def _is_bullet_marker_text(text: str) -> bool:
-    return bool(BULLET_PREFIX_RE.match(f"{str(text or '').strip()} x"))
-
-
-def _ends_sentence(line: str) -> bool:
-    return bool(re.search(r"[.!?;:。！？]$" , str(line or "").strip()))
-
-
-def _is_body_heading_line(line: str) -> bool:
-    text = str(line or "").strip()
-    if not text:
-        return False
-    return bool(re.match(r"^(?:chapter|section|appendix)\b", text, flags=re.IGNORECASE))
-
-
-def _line_kind(line: dict) -> str:
-    text = str(line.get("text") or "").strip()
-    if _is_body_heading_line(text):
-        return "heading"
-    return "paragraph"
-
-
-def _style_signature(line: dict) -> tuple:
-    color = line.get("color")
-    normalized_color = None
-    if isinstance(color, tuple):
-        normalized_color = tuple(round(float(value), 3) for value in color[:3])
-    return (
-        str(line.get("fontname") or ""),
-        bool(line.get("is_bold")),
-        bool(line.get("is_italic")),
-        normalized_color,
-    )
-
-
-def _is_list_continuation_line(line: dict, previous: dict, anchor_x: float) -> bool:
-    text = str(line.get("text") or "").strip()
-    if not text or bool(line.get("marker_candidate")) or _is_bullet_line(text) or _is_body_heading_line(text):
-        return False
-
-    line_gap = float(line.get("top", 0.0)) - float(previous.get("bottom", 0.0))
-    gap_close = line_gap <= max(6.0, float(previous.get("size", 0.0)) * 0.9)
-    size_close = abs(float(line.get("size", 0.0)) - float(previous.get("size", 0.0))) <= 0.8
-    style_close = _style_signature(line) == _style_signature(previous)
-    indent_x = float(line.get("x0", 0.0))
-    aligned_with_text = abs(indent_x - anchor_x) <= 8.0
-    return gap_close and size_close and style_close and aligned_with_text
-
-
-def _looks_like_inline_term_continuation(line: dict) -> bool:
-    text = str(line.get("text") or "").strip()
-    if not text:
-        return False
-    tokens = text.split()
-    if len(tokens) == 1 and not _ends_sentence(text):
-        return True
-    return bool(line.get("has_mixed_styles")) and int(line.get("word_count", 0)) >= 2
-
-
-def _has_room_for_next_line_start(previous: dict, line: dict) -> bool:
-    body_right = float(previous.get("body_right", previous.get("x1", 0.0)))
-    remaining_width = body_right - float(previous.get("x1", 0.0))
-    first_word_width = float(line.get("first_word_width", 0.0))
-    if first_word_width <= 0.0:
-        return False
-    return remaining_width >= first_word_width * 1.25
-
-
-def _should_merge_paragraph_lines(previous: dict, line: dict) -> bool:
-    line_gap = float(line.get("top", 0.0)) - float(previous.get("bottom", 0.0))
-    return line_gap <= 5.0
-
-
-def _normalize_list_block_lines(lines: Sequence[dict]) -> List[str]:
-    normalized: List[str] = []
-    current_item: str | None = None
-    current_depth = 0
-    current_text_start_x: float | None = None
-    marker_positions = sorted(
-        {
-            round(float(line.get("marker_x", line.get("x0", 0.0))), 2)
-            for line in lines
-            if bool(line.get("marker_candidate")) or _is_bullet_line(str(line.get("text") or "").strip())
-        }
-    )
-
-    def _item_prefix(depth: int) -> str:
-        markers = ["-", "*", "+"]
-        return f"{'  ' * depth}{markers[depth % len(markers)]} "
-
-    def _continuation_prefix(depth: int) -> str:
-        return f"{'  ' * (depth + 1)}"
-
-    def _strip_marker_text(line: dict) -> str:
-        text = str(line.get("text") or "").strip()
-        if not text:
-            return text
-        parts = text.split(maxsplit=1)
-        if len(parts) == 1:
-            return ""
-        first = parts[0]
-        if bool(line.get("marker_candidate")) or _is_bullet_marker_text(first) or _is_bullet_line(text):
-            return parts[1].strip()
-        return text
-
-    for line in lines:
-        text = str(line.get("text") or "").strip()
-        if not text:
-            continue
-        if bool(line.get("marker_candidate")) or _is_bullet_line(text):
-            if current_item:
-                normalized.append(current_item)
-            marker_x = round(float(line.get("marker_x", line.get("x0", 0.0))), 2)
-            try:
-                current_depth = marker_positions.index(marker_x)
-            except ValueError:
-                current_depth = 0
-            current_text_start_x = float(line.get("text_start_x", line.get("x0", 0.0)))
-            current_item = f"{_item_prefix(current_depth)}{_strip_marker_text(line)}".rstrip()
-            continue
-        if current_item and current_item.endswith("-"):
-            current_item = f"{current_item}{text}".strip()
-            continue
-        if current_item:
-            line_text_start_x = float(line.get("text_start_x", line.get("x0", 0.0)))
-            if current_text_start_x is not None and abs(line_text_start_x - current_text_start_x) <= 8.0:
-                current_item = f"{current_item} {text}".strip()
-                continue
-            normalized.append(current_item)
-            current_item = None
-            normalized.append(f"{_continuation_prefix(current_depth)}{text}".rstrip())
-            continue
-        normalized.append(text)
-
-    if current_item:
-        normalized.append(current_item)
-
-    return normalized
-
-
-def _build_body_blocks(lines: Sequence[dict]) -> List[dict]:
-    if not lines:
-        return []
-
-    blocks: List[dict] = []
-    current_block: dict | None = None
-
-    for line in lines:
-        kind = _line_kind(line)
-        if current_block is None:
-            current_block = {
-                "kind": kind,
-                "lines": [line],
-            }
-            continue
-
-        previous = current_block["lines"][-1]
-        same_kind = current_block["kind"] == kind
-
-        if same_kind and kind == "paragraph" and _should_merge_paragraph_lines(previous, line):
-            current_block["lines"].append(line)
-            continue
-
-        blocks.append(current_block)
-        current_block = {
-            "kind": kind,
-            "lines": [line],
-        }
-
-    if current_block is not None:
-        blocks.append(current_block)
-
-    return blocks
-
-
-def _extract_body_word_lines(
-    page: "pdfplumber.page.Page",
-    header_margin: float,
-    footer_margin: float,
-    excluded_bboxes: Sequence[Tuple[float, float, float, float]] = (),
-) -> List[dict]:
-    filtered_page = _filter_page_for_extraction(page)
-    body_top, body_bottom = _detect_body_bounds(page, header_margin=header_margin, footer_margin=footer_margin)
-    words = filtered_page.extract_words(
-        x_tolerance=1.5,
-        y_tolerance=2.0,
-        keep_blank_chars=False,
-        extra_attrs=["size", "fontname"],
-    ) or []
-
-    def _word_bbox(word: dict) -> Tuple[float, float, float, float]:
-        top = float(word.get("top", 0.0))
-        bottom = float(word.get("bottom", top))
-        return (
-            float(word.get("x0", 0.0)),
-            top,
-            float(word.get("x1", 0.0)),
-            bottom,
-        )
-
-    def _word_in_body(word: dict) -> bool:
-        bbox = _word_bbox(word)
-        if bbox[3] <= body_top or bbox[1] >= body_bottom:
-            return False
-        return not any(_bboxes_intersect(bbox, excluded_bbox) for excluded_bbox in excluded_bboxes)
-
-    filtered_words = [word for word in words if _word_in_body(word)]
-    grouped_lines: List[List[dict]] = []
-    for word in sorted(filtered_words, key=lambda item: (float(item.get("top", 0.0)), float(item.get("x0", 0.0)))):
-        cleaned = _repair_watermark_bleed(str(word.get("text") or "").strip())
-        if not cleaned or _is_layout_artifact(cleaned):
-            continue
-        if not grouped_lines or abs(float(word.get("top", 0.0)) - float(grouped_lines[-1][0].get("top", 0.0))) > 2.5:
-            grouped_lines.append([word])
-            continue
-        grouped_lines[-1].append(word)
-
-    lines: List[dict] = []
-    for words_in_line in grouped_lines:
-        ordered = sorted(words_in_line, key=lambda item: float(item.get("x0", 0.0)))
-        text_parts = []
-        for word in ordered:
-            cleaned = _repair_watermark_bleed(str(word.get("text") or "").strip())
-            if cleaned:
-                text_parts.append(cleaned)
-        text = " ".join(text_parts).strip()
-        if not text or _is_layout_artifact(text):
-            continue
-        fontnames = [str(word.get("fontname") or "") for word in ordered if str(word.get("fontname") or "")]
-        dominant_font = max(fontnames, key=fontnames.count) if fontnames else ""
-        colors = [word.get("non_stroking_color") or word.get("stroking_color") for word in ordered]
-        normalized_colors = [color for color in colors if isinstance(color, tuple) and len(color) >= 3]
-        dominant_color = None
-        if normalized_colors:
-            color_keys = [tuple(round(float(value), 3) for value in color[:3]) for color in normalized_colors]
-            dominant_color = max(color_keys, key=color_keys.count)
-        first_cleaned = _repair_watermark_bleed(str(ordered[0].get("text") or "").strip())
-        second_word = ordered[1] if len(ordered) > 1 else ordered[0]
-        marker_gap = float(second_word.get("x0", ordered[0].get("x1", 0.0))) - float(ordered[0].get("x1", 0.0))
-        marker_candidate = len(ordered) > 1 and (
-            _is_bullet_marker_text(first_cleaned)
-            or (
-                len(first_cleaned) == 1
-                and not first_cleaned.isalnum()
-                and marker_gap >= 4.0
-            )
-            or (
-                first_cleaned in {"o", "O", "?", "\uFFFD"}
-                and marker_gap >= 4.0
-            )
-        )
-        first_non_bullet_word = next(
-            (
-                word
-                for word in ordered
-                if not marker_candidate
-                or not _is_bullet_marker_text(_repair_watermark_bleed(str(word.get("text") or "").strip()))
-                and _repair_watermark_bleed(str(word.get("text") or "").strip()) not in {"o", "O", "?", "\uFFFD"}
-            ),
-            ordered[0],
-        )
-        word_style_signatures = []
-        for word in ordered:
-            fontname = str(word.get("fontname") or "")
-            color = word.get("non_stroking_color") or word.get("stroking_color")
-            normalized_color = None
-            if isinstance(color, tuple) and len(color) >= 3:
-                normalized_color = tuple(round(float(value), 3) for value in color[:3])
-            word_style_signatures.append(
-                (
-                    fontname,
-                    bool(re.search(r"bold", fontname, flags=re.IGNORECASE)),
-                    bool(re.search(r"(italic|oblique)", fontname, flags=re.IGNORECASE)),
-                    normalized_color,
-                )
-            )
-        lines.append(
-            {
-                "text": text,
-                "x0": float(ordered[0].get("x0", 0.0)),
-                "x1": float(ordered[-1].get("x1", 0.0)),
-                "top": min(float(word.get("top", 0.0)) for word in ordered),
-                "bottom": max(float(word.get("bottom", 0.0)) for word in ordered),
-                "size": sum(float(word.get("size", 0.0)) for word in ordered) / max(len(ordered), 1),
-                "fontname": dominant_font,
-                "color": dominant_color,
-                "is_bold": bool(re.search(r"bold", dominant_font, flags=re.IGNORECASE)),
-                "is_italic": bool(re.search(r"(italic|oblique)", dominant_font, flags=re.IGNORECASE)),
-                "marker_candidate": marker_candidate,
-                "text_start_x": float(first_non_bullet_word.get("x0", ordered[0].get("x0", 0.0))),
-                "first_word_width": float(first_non_bullet_word.get("x1", 0.0)) - float(first_non_bullet_word.get("x0", 0.0)),
-                "body_right": float(getattr(page, "width", 0.0)),
-                "word_count": len(ordered),
-                "has_mixed_styles": len(set(word_style_signatures)) > 1,
-                "first_word_style_signature": word_style_signatures[0] if word_style_signatures else None,
-            }
-        )
-
-    return lines
-
-
-def _normalize_body_lines(lines: Sequence[str]) -> List[str]:
-    normalized: List[str] = []
-    buffer: List[str] = []
-
-    def _flush_buffer() -> None:
-        nonlocal buffer
-        if buffer:
-            normalized.append(" ".join(buffer))
-            buffer = []
-
-    for raw_line in lines:
-        line = str(raw_line or "").strip()
-        if not line:
-            continue
-        if _is_bullet_line(line) or _is_body_heading_line(line):
-            _flush_buffer()
-            normalized.append(line)
-            continue
-        if buffer and str(buffer[-1]).endswith("-"):
-            buffer[-1] = f"{buffer[-1]}{line}".strip()
-            continue
-        if buffer and _ends_sentence(buffer[-1]):
-            _flush_buffer()
-        buffer.append(line)
-
-    _flush_buffer()
-    return normalized
-
-
-def _join_non_heading_block_lines(lines: Sequence[str]) -> str:
-    joined = [str(raw_line or "").strip() for raw_line in lines if str(raw_line or "").strip()]
-    return " ".join(joined).strip()
-
-
-def _normalize_cell_lines(cell: str) -> List[str]:
-    raw_lines = [part.strip() for part in str(cell or "").splitlines()]
-    lines = _remove_watermark_fragment_lines(raw_lines)
-    if not lines:
-        return []
-
-    logical_lines: List[str] = []
-    buffer: List[str] = []
-
-    def _flush_buffer() -> None:
-        nonlocal buffer
-        if buffer:
-            logical_lines.append(" ".join(buffer))
-            buffer = []
-
-    for line in lines:
-        if _is_bullet_line(line):
-            _flush_buffer()
-            logical_lines.append(line)
-            continue
-        if logical_lines and _is_bullet_line(logical_lines[-1]) and not buffer:
-            logical_lines[-1] = f"{logical_lines[-1]} {line}".strip()
-            continue
-        if buffer and str(buffer[-1]).endswith("-"):
-            buffer[-1] = f"{buffer[-1]}{line}".strip()
-            continue
-        if buffer and _ends_sentence(buffer[-1]):
-            _flush_buffer()
-        buffer.append(line)
-
-    _flush_buffer()
-    return logical_lines
-
-
-def _collapse_structural_triplet_columns(table: Sequence[Sequence[str]]) -> List[List[str]]:
-    rows = [list(row) for row in table]
-    if not rows:
-        return []
-
-    col_count = max((len(row) for row in rows), default=0)
-    if col_count == 0 or col_count % 3 != 0:
-        return [list(row) for row in rows]
-
-    padded_rows = [list(row) + [""] * (col_count - len(row)) for row in rows]
-    collapsed_indices: List[int] = []
-
-    for start in range(0, col_count, 3):
-        left_values = [_normalize_text(row[start]) for row in padded_rows]
-        right_values = [_normalize_text(row[start + 2]) for row in padded_rows]
-        if any(left_values) or any(right_values):
-            collapsed_indices.extend([start, start + 1, start + 2])
-            continue
-        collapsed_indices.append(start + 1)
-
-    return [[row[idx] for idx in collapsed_indices] for row in padded_rows]
-
-
-def _normalize_extracted_table(table: Sequence[Sequence[str]]) -> List[List[str]]:
-    # TODO: Extend structural-column collapse beyond strict 3-column spacer
-    # triples when real documents expose asymmetric spacer patterns.
-    normalized: List[List[str]] = []
-    for row in table:
-        normalized_row = []
-        for cell in row:
-            normalized_row.append("\n".join(_normalize_cell_lines(str(cell or ""))))
-        normalized.append(normalized_row)
-    return _collapse_structural_triplet_columns(normalized)
-
-
-def _looks_like_table(table: Sequence[Sequence[str]]) -> bool:
-    return _table_rejection_reason(table) is None
-
-
-def _table_rejection_reason(table: Sequence[Sequence[str]]) -> str | None:
-    if not table:
-        return "empty table"
-
-    normalized_rows = [[str(cell or "").strip() for cell in row] for row in table]
-
-    if not any(cell for cell in normalized_rows[0]):
-        return "empty first row"
-
-    return None
-
-
-def _log_rejected_table(
-    table: Sequence[Sequence[str]],
-    crop_bbox: Tuple[float, float, float, float],
-    reason: str,
-) -> None:
-    row_count = len(table)
-    col_count = max((len(row) for row in table), default=0)
-    bbox_text = ", ".join(f"{value:.2f}" for value in crop_bbox)
-    print(
-        f"[table-reject] bbox=({bbox_text}) rows={row_count} cols={col_count} reason={reason}"
-    )
-
-
-def _looks_like_header_row(row: Sequence[str]) -> bool:
-    if not row:
-        return False
-
-    normalized = [_normalize_text(c) for c in row]
-    tokens = [cell for cell in normalized if cell]
-    if not tokens:
-        return False
-
-    alpha_like = sum(1 for token in tokens if re.fullmatch(r"[A-Za-z][A-Za-z0-9\s/&._:-]*", token))
-    short = sum(1 for token in tokens if len(token) <= 24)
-
-    return alpha_like >= len(tokens) * 0.8 and short >= len(tokens) * 0.8
-
-
-def _rows_match(a: Sequence[str], b: Sequence[str]) -> bool:
-    if len(a) != len(b):
-        return False
-    return all(_normalize_text(x) == _normalize_text(y) for x, y in zip(a, b))
-
-
-def _split_repeated_header(prev_rows: TableRows, curr_rows: TableRows) -> TableRows:
-    if prev_rows and curr_rows and _rows_match(prev_rows[0], curr_rows[0]):
-        return curr_rows[1:]
-    return curr_rows
-
-
-def _is_continuation_chunk(prev_rows: TableRows, curr_rows: TableRows) -> bool:
-    if not prev_rows or not curr_rows:
-        return False
-    if len(prev_rows[0]) != len(curr_rows[0]):
-        return False
-
-    # Continuation fragments are usually headerless and keep first column blank while body continues.
-    first = curr_rows[0]
-    if not first:
-        return False
-    if _looks_like_header_row(first):
-        return False
-    if _normalize_text(first[0]):
-        return False
-
-    return any(_normalize_text(cell) for cell in first[1:])
-
-
-def _extract_continuation_lines(
-    page_text: str,
-    repeated_header: str,
-    next_row_label: str,
-) -> List[str]:
-    lines = [line.strip() for line in str(page_text or "").splitlines() if line.strip()]
-    if not lines:
-        return []
-
-    try:
-        start_idx = next(i for i, line in enumerate(lines) if _normalize_text(line) == _normalize_text(repeated_header))
-    except StopIteration:
-        return []
-
-    tail = lines[start_idx + 1 :]
-    if not tail:
-        return []
-
-    stop_idx = len(tail)
-    normalized_label = _normalize_text(next_row_label)
-    for i, line in enumerate(tail):
-        if _normalize_text(line).startswith(normalized_label):
-            stop_idx = i
-            break
-    return tail[:stop_idx]
-
-
-def _normalize_two_col_continuation_rows(rows: TableRows) -> TableRows:
-    return [["", str(row[0] or "").strip(), str(row[1] or "").strip()] for row in rows if row]
-
-
-def _should_try_table_continuation_merge(
-    pending_page: int | None,
-    current_page: int,
-) -> bool:
-    return pending_page is not None and current_page == pending_page + 1
-
-
-def _bboxes_intersect(
-    a: Tuple[float, float, float, float],
-    b: Tuple[float, float, float, float],
-) -> bool:
-    ax0, ay0, ax1, ay1 = a
-    bx0, by0, bx1, by1 = b
-    return ax0 < bx1 and ax1 > bx0 and ay0 < by1 and ay1 > by0
-
-
-def _body_text_boxes(
-    page: pdfplumber.page.PageObject,
-    header_margin: float,
-    footer_margin: float,
-    excluded_bboxes: Sequence[Tuple[float, float, float, float]] = (),
-) -> List[Tuple[float, float, float, float]]:
-    filtered_page = _filter_page_for_extraction(page)
-    body_top, body_bottom = _detect_body_bounds(
-        page,
-        header_margin=header_margin,
-        footer_margin=footer_margin,
-    )
-    text_boxes: List[Tuple[float, float, float, float]] = []
-    for word in filtered_page.extract_words() or []:
-        text = _normalize_text(str(word.get("text") or ""))
-        if not text or _is_layout_artifact(text):
-            continue
-        bbox = (
-            float(word.get("x0", 0.0)),
-            float(word.get("top", 0.0)),
-            float(word.get("x1", 0.0)),
-            float(word.get("bottom", 0.0)),
-        )
-        if bbox[3] <= body_top or bbox[1] >= body_bottom:
-            continue
-        if any(_bboxes_intersect(bbox, excluded_bbox) for excluded_bbox in excluded_bboxes):
-            continue
-        text_boxes.append(bbox)
-    return text_boxes
-
-
-def _gap_text_boxes_after_bbox(
-    page: pdfplumber.page.PageObject,
-    bbox: Tuple[float, float, float, float],
-    table_bboxes: Sequence[Tuple[float, float, float, float]],
-    header_margin: float,
-    footer_margin: float,
-) -> List[Tuple[float, float, float, float]]:
-    return [
-        text_bbox
-        for text_bbox in _body_text_boxes(
-            page,
-            header_margin=header_margin,
-            footer_margin=footer_margin,
-            excluded_bboxes=table_bboxes,
-        )
-        if float(text_bbox[1]) >= float(bbox[3])
-    ]
-
-
-def _gap_text_boxes_before_bbox(
-    page: pdfplumber.page.PageObject,
-    bbox: Tuple[float, float, float, float],
-    table_bboxes: Sequence[Tuple[float, float, float, float]],
-    header_margin: float,
-    footer_margin: float,
-) -> List[Tuple[float, float, float, float]]:
-    return [
-        text_bbox
-        for text_bbox in _body_text_boxes(
-            page,
-            header_margin=header_margin,
-            footer_margin=footer_margin,
-            excluded_bboxes=table_bboxes,
-        )
-        if float(text_bbox[3]) <= float(bbox[1])
-    ]
-
-
-def _vertical_axes_for_bbox(
-    page: pdfplumber.page.PageObject,
-    bbox: Tuple[float, float, float, float],
-) -> List[float]:
-    x0, y0, x1, y1 = bbox
-    axes = [
-        float(edge["x0"])
-        for edge in page.vertical_edges
-        if float(edge["x0"]) >= x0 - 2.0
-        and float(edge["x0"]) <= x1 + 2.0
-        and float(edge["bottom"]) >= y0
-        and float(edge["top"]) <= y1
-    ]
-    return _merge_numeric_positions(axes, tolerance=1.0)
-
-
-def _continuation_regions_should_merge(
-    prev_bbox: Tuple[float, float, float, float],
-    curr_bbox: Tuple[float, float, float, float],
-    prev_axes: Sequence[float],
-    curr_axes: Sequence[float],
-    body_top: float,
-    body_bottom: float,
-    gap_text_boxes: Sequence[Tuple[float, float, float, float]],
-    edge_tolerance: float = 24.0,
-    axis_tolerance: float = 1.0,
-) -> bool:
-    _prev_x0, prev_top, _prev_x1, prev_bottom = prev_bbox
-    _curr_x0, curr_top, _curr_x1, curr_bottom = curr_bbox
-
-    shared_axes = [
-        axis
-        for axis in prev_axes
-        if any(abs(axis - other) <= axis_tolerance for other in curr_axes)
-    ]
-    if not shared_axes:
-        return False
-
-    if gap_text_boxes:
-        return False
-
-    prev_near_footer = abs(body_bottom - prev_bottom) <= edge_tolerance
-    curr_near_header = abs(curr_top - body_top) <= edge_tolerance
-    if prev_near_footer or curr_near_header:
-        return True
-
-    return True
-
-
-def _maybe_merge_missing_first_column_chunk(
-    pending_table: TableRows | None,
-    current_rows: TableRows,
-    page_text: str,
-) -> TableRows | None:
-    if not pending_table or not current_rows:
-        return None
-    if len(pending_table[0]) != 3:
-        return None
-    if any(len(row) != 2 for row in current_rows):
-        return None
-
-    last_row = pending_table[-1]
-    if len(last_row) < 3 or not _normalize_text(last_row[1]):
-        return None
-
-    repeated_header = " ".join(str(cell or "").strip() for cell in pending_table[0]).strip()
-    continuation_lines = _extract_continuation_lines(page_text, repeated_header, str(current_rows[0][0] or ""))
-    if continuation_lines:
-        joiner = "\n" if str(last_row[2] or "").strip() else ""
-        last_row[2] = f"{last_row[2]}{joiner}" + "\n".join(continuation_lines)
-
-    normalized_rows = _normalize_two_col_continuation_rows(current_rows)
-    if not normalized_rows:
-        return None
-
-    pending_table.extend(normalized_rows)
-    return pending_table
-
-
-def _format_markdown_cell(value: str) -> str:
-    lines = _normalize_cell_lines(value)
-    if not lines:
-        return ""
-    return "<br>".join(line.replace("|", "\\|") for line in lines)
-
-
-def _table_regions(
-    page: pdfplumber.page.PageObject,
-    y_tolerance: float = 65.0,
-    min_lines: int = 3,
-) -> List[tuple]:
-    del y_tolerance  # table selection now relies on edge connectivity, not vertical-gap grouping
-
-    body_top, body_bottom = _detect_body_bounds(page, header_margin=90.0, footer_margin=40.0)
-
-    horizontal_edges = [
-        edge
-        for edge in page.horizontal_edges
-        if float(edge["x1"]) - float(edge["x0"]) >= 50.0
-        and float(edge["bottom"]) > body_top
-        and float(edge["top"]) < body_bottom
-    ]
-    vertical_edges = [
-        edge
-        for edge in page.vertical_edges
-        if float(edge["bottom"]) > body_top
-        and float(edge["top"]) < body_bottom
-    ]
-
-    merged_h = []
-    for group in _build_segment_groups(
-        horizontal_edges,
-        axis_key="top",
-        merge_fn=_merge_horizontal_band_segments,
-        tolerance=1.0,
-    ):
-        merged_h.extend(group["merged_segments"])
-
-    merged_v = []
-    for group in _build_segment_groups(
-        vertical_edges,
-        axis_key="x0",
-        merge_fn=_merge_vertical_band_segments,
-        tolerance=1.0,
-    ):
-        merged_v.extend(group["merged_segments"])
-
-    if not merged_h:
-        return []
-
-    graph: List[set[int]] = [set() for _ in range(len(merged_h))]
-    component_verticals: List[set[int]] = [set() for _ in range(len(merged_h))]
-    tolerance = 1.0
-
-    for h_idx, h_edge in enumerate(merged_h):
-        for v_idx, v_edge in enumerate(merged_v):
-            intersects = (
-                float(v_edge["x0"]) >= float(h_edge["x0"]) - tolerance
-                and float(v_edge["x0"]) <= float(h_edge["x1"]) + tolerance
-                and float(h_edge["top"]) >= float(v_edge["top"]) - tolerance
-                and float(h_edge["top"]) <= float(v_edge["bottom"]) + tolerance
-            )
-            if not intersects:
-                continue
-            component_verticals[h_idx].add(v_idx)
-
-    for i in range(len(merged_h)):
-        for j in range(i + 1, len(merged_h)):
-            if component_verticals[i] and component_verticals[j]:
-                shared_vertical = component_verticals[i].intersection(component_verticals[j])
-                if shared_vertical:
-                    graph[i].add(j)
-                    graph[j].add(i)
-
-    visited = set()
-    groups: List[tuple] = []
-    for start in range(len(merged_h)):
-        if start in visited:
-            continue
-        stack = [start]
-        component = []
-        shared_verticals = set()
-        while stack:
-            idx = stack.pop()
-            if idx in visited:
-                continue
-            visited.add(idx)
-            component.append(idx)
-            shared_verticals.update(component_verticals[idx])
-            stack.extend(graph[idx] - visited)
-
-        component_lines = [merged_h[idx] for idx in component]
-        if len(component_lines) < min_lines:
-            continue
-        if not shared_verticals:
-            continue
-
-        x0 = min(float(edge["x0"]) for edge in component_lines)
-        x1 = max(float(edge["x1"]) for edge in component_lines)
-        if shared_verticals:
-            x0 = min(x0, *(float(merged_v[idx]["x0"]) for idx in shared_verticals))
-            x1 = max(x1, *(float(merged_v[idx]["x1"]) for idx in shared_verticals))
-        groups.append((x0, x1, component_lines))
-
-    return sorted(groups, key=lambda item: min(float(edge["top"]) for edge in item[2]))
-
-
-def _extract_tables_from_crop(
-    page: pdfplumber.page.PageObject,
-    crop_bbox: Tuple[float, float, float, float],
-) -> List[TableChunk]:
-    x0, y0, x1, y1 = crop_bbox
-    crop = page.crop(crop_bbox)
-
-    # Keep left/right outer lines to avoid merged adjacent-table results.
-    y_min, y_max = y0, y1
-    v_lines = []
-    for edge in page.vertical_edges:
-        if edge["x0"] < x0 or edge["x0"] > x1:
-            continue
-        if edge["top"] > y_max or edge["bottom"] < y_min:
-            continue
-        v_lines.append(edge["x0"])
-
-    explicit_v = sorted({x0, x1, *v_lines})
-
-    candidates = [
-        {
-            "vertical_strategy": "lines",
-            "horizontal_strategy": "lines",
-            "explicit_vertical_lines": explicit_v,
-            "snap_tolerance": 3,
-            "join_tolerance": 3,
-            "intersection_tolerance": 2,
-            "min_words_vertical": 1,
-            "min_words_horizontal": 1,
-        },
-        {
-            "vertical_strategy": "text",
-            "horizontal_strategy": "lines",
-            "explicit_vertical_lines": explicit_v,
-            "snap_tolerance": 4,
-            "join_tolerance": 4,
-            "intersection_tolerance": 2,
-            "min_words_vertical": 1,
-            "min_words_horizontal": 1,
-        },
-    ]
-
-    for settings in candidates:
-        tables = crop.extract_tables(table_settings=settings) or []
-        cleaned = []
-        for table in tables:
-            reason = _table_rejection_reason(table)
-            if reason is not None:
-                _log_rejected_table(table, crop_bbox, reason)
-                continue
-            cleaned.append(_merge_cells(table))
-        if cleaned:
-            return [(table, crop_bbox) for table in cleaned]
-    return []
-
-
-def _extract_tables(
-    page: pdfplumber.page.PageObject,
-    force_table: bool = False,
-) -> List[TableChunk]:
-    page = _filter_page_for_extraction(page)
-    seen_keys = set()
-    merged: List[TableChunk] = []
-    # Targeted extraction from table-like regions with missing outer vertical
-    # borders. This is preferred for docs without full edge lines.
-    table_regions = _table_regions(page)
-    for x0, x1, lines in table_regions:
-        y0 = min(edge["top"] for edge in lines) - 2
-        y1 = max(edge["top"] for edge in lines) + 2
-        crop_bbox = (
-            max(0.0, x0),
-            max(0.0, y0),
-            min(page.width, x1),
-            min(page.height, y1),
-        )
-        for table, crop_box in _extract_tables_from_crop(page, crop_bbox):
-            table = _normalize_extracted_table(table)
-            rows_key = tuple(tuple(row) for row in table)
-            bbox_key = tuple(round(v, 2) for v in crop_box)
-            key = (rows_key, bbox_key)
-            if key not in seen_keys:
-                seen_keys.add(key)
-                merged.append((table, crop_box))
-
-    if merged:
-        return merged
-
-    if not force_table:
-        return merged
-
-    # Fallback to page-wide extraction when region-based cues are unavailable
-    # and the caller explicitly requests aggressive table extraction.
-    full_bbox = (0.0, 0.0, float(page.width), float(page.height))
-    fallback_settings = [
-        {
-            "vertical_strategy": "lines",
-            "horizontal_strategy": "lines",
-            "snap_tolerance": 3,
-            "join_tolerance": 3,
-            "intersection_tolerance": 2,
-            "min_words_vertical": 1,
-            "min_words_horizontal": 1,
-        },
-        {
-            "vertical_strategy": "text",
-            "horizontal_strategy": "lines",
-            "snap_tolerance": 4,
-            "join_tolerance": 4,
-            "intersection_tolerance": 2,
-            "min_words_vertical": 1,
-            "min_words_horizontal": 1,
-        },
-        {
-            "vertical_strategy": "text",
-            "horizontal_strategy": "text",
-            "snap_tolerance": 4,
-            "join_tolerance": 4,
-            "intersection_tolerance": 2,
-            "min_words_vertical": 1,
-            "min_words_horizontal": 1,
-        },
-    ]
-
-    for settings in fallback_settings:
-        tables = page.extract_tables(table_settings=settings) or []
-        cleaned = []
-        for table in tables:
-            reason = _table_rejection_reason(table)
-            if reason is not None:
-                _log_rejected_table(table, full_bbox, reason)
-                continue
-            cleaned.append(_merge_cells(table))
-        if cleaned:
-            for table in cleaned:
-                table = _normalize_extracted_table(table)
-                rows_key = tuple(tuple(row) for row in table)
-                bbox_key = tuple(round(v, 2) for v in full_bbox)
-                key = (rows_key, bbox_key)
-                if key not in seen_keys:
-                    seen_keys.add(key)
-                    merged.append((table, full_bbox))
-            break
-
-    return merged
-
-
-def _table_text_from_rows(rows: Sequence[Sequence[str]]) -> str:
-    if not rows:
-        return ""
-
-    header = [str(col or "").strip() for col in rows[0]]
-    body = rows[1:]
-
-    if not body:
-        body = rows
-        header = [f"Column {idx}" for idx in range(1, len(rows[0]) + 1)]
-
-    header_line = "| " + " | ".join(cell or f"Column {idx + 1}" for idx, cell in enumerate(header)) + " |"
-    divider_line = "| " + " | ".join("---" for _ in header) + " |"
-    body_lines = []
-    for row in body:
-        padded_row = list(row) + [""] * max(0, len(header) - len(row))
-        body_lines.append("| " + " | ".join(_format_markdown_cell(str(value or "")) for value in padded_row) + " |")
-
-    return "\n".join([header_line, divider_line, *body_lines])
-
-
-def _merge_split_rows(rows: TableRows) -> TableRows:
-    if not rows:
-        return rows
-
-    merged: TableRows = [list(rows[0])]
-    for row in rows[1:]:
-        non_empty = [idx for idx, cell in enumerate(row) if _normalize_text(cell)]
-        if len(merged) > 1 and row and not _normalize_text(row[0]):
-            previous = merged[-1]
-            previous_second = _normalize_text(previous[1]) if len(previous) > 1 else ""
-            current_second = _normalize_text(row[1]) if len(row) > 1 else ""
-
-            if len(non_empty) == 1 and non_empty[0] > 0:
-                idx = non_empty[0]
-                joiner = "\n" if previous[idx].strip() else ""
-                previous[idx] = f"{previous[idx]}{joiner}{row[idx]}".strip()
-                continue
-
-            if (
-                len(non_empty) >= 2
-                and 1 in non_empty
-                and 2 in non_empty
-                and previous_second
-                and current_second == previous_second
-            ):
-                joiner = "\n" if previous[2].strip() else ""
-                previous[2] = f"{previous[2]}{joiner}{row[2]}".strip()
-                continue
-        merged.append(list(row))
-    return merged
-
-
-def _append_output_table(output_tables: List[str], page_no: int, table_no: int, table_rows: TableRows) -> None:
-    table_text = _table_text_from_rows(_merge_split_rows(table_rows))
-    if table_text:
-        output_tables.append(f"### Page {page_no} table {table_no}\n{table_text}")
-
-
-def _body_excluded_bboxes(
-    pending_table: Optional[TableRows],
-    tables: Sequence[TableChunk],
-    body_top: float,
-) -> List[Tuple[float, float, float, float]]:
-    excluded = [bbox for _rows, bbox in tables]
-    if not pending_table or not tables:
-        return excluded
-
-    first_rows, first_bbox = tables[0]
-    if len(pending_table[0]) == 3 and first_rows and all(len(row) == 2 for row in first_rows):
-        x0, _top, x1, bottom = first_bbox
-        excluded[0] = (x0, body_top, x1, bottom)
-    return excluded
-
-
-def _image_intersects_body(
-    image_meta: dict,
-    body_top: float,
-    body_bottom: float,
-) -> bool:
-    top = float(image_meta.get("top", 0.0))
-    bottom = float(image_meta.get("bottom", top))
-    return bottom > body_top and top < body_bottom
-
-
-def _extract_embedded_images(
-    pdf_path: Path,
-    out_image_dir: Path,
-    stem: str,
-    pages: Optional[Sequence[int]] = None,
-) -> List[Path]:
-    out_image_dir.mkdir(parents=True, exist_ok=True)
-
-    image_files: List[Path] = []
-    selected_pages = set(int(page_no) for page_no in (pages or []))
-    reader = PdfReader(str(pdf_path))
-    with pdfplumber.open(str(pdf_path)) as plumber_pdf:
-        for page_idx, (page, plumber_page) in enumerate(zip(reader.pages, plumber_pdf.pages), start=1):
-            if selected_pages and page_idx not in selected_pages:
-                continue
-
-            body_top, body_bottom = _detect_body_bounds(
-                plumber_page,
-                header_margin=90.0,
-                footer_margin=40.0,
-            )
-            allowed_names = {
-                str(image_meta.get("name") or "")
-                for image_meta in plumber_page.images
-                if _image_intersects_body(image_meta, body_top=body_top, body_bottom=body_bottom)
-            }
-
-            kept_idx = 0
-            for image_file in page.images:
-                image_name = Path(image_file.name or "").name
-                image_stem = Path(image_name).stem
-                if image_stem not in allowed_names and image_name not in allowed_names:
-                    continue
-                kept_idx += 1
-                suffix = Path(image_name).suffix or ".bin"
-                out_path = out_image_dir / f"{stem}_page_{page_idx:02d}_image_{kept_idx:02d}{suffix}"
-                out_path.write_bytes(image_file.data)
-                image_files.append(out_path)
-
-    return image_files
-
-
-def extract_pdf_to_outputs(
-    pdf_path: Path,
-    out_md_dir: Path,
-    out_image_dir: Path,
-    stem: str,
-    header_margin: float = 90,
-    footer_margin: float = 40,
-    pages: Optional[Sequence[int]] = None,
-    force_table: bool = False,
-    debug: bool = False,
-    debug_watermark: bool = False,
-) -> dict:
-    out_md_dir.mkdir(parents=True, exist_ok=True)
-
-    output_text = []
-    output_tables = []
-    table_debug_pages: List[dict] = []
-    edge_debug_pages: List[dict] = []
-    rotated_debug: List[dict] = []
-    selected_pages = set(int(page_no) for page_no in (pages or []))
-
-    pending_table: Optional[TableRows] = None
-    pending_page: Optional[int] = None
-    pending_last_page: Optional[int] = None
-    pending_bbox: Optional[Tuple[float, float, float, float]] = None
-    pending_axes: List[float] = []
-    pending_gap_text_boxes: List[Tuple[float, float, float, float]] = []
-
-    def _flush_pending() -> None:
-        nonlocal pending_table, pending_page, pending_last_page, pending_bbox, pending_axes, pending_gap_text_boxes
-        if pending_table is not None and pending_page is not None:
-            _append_output_table(output_tables, pending_page, len(output_tables) + 1, pending_table)
-        pending_table = None
-        pending_page = None
-        pending_last_page = None
-        pending_bbox = None
-        pending_axes = []
-        pending_gap_text_boxes = []
-
-    with pdfplumber.open(str(pdf_path)) as pdf:
-        for page_idx, page in enumerate(pdf.pages, start=1):
-            if selected_pages and page_idx not in selected_pages:
-                _flush_pending()
-                continue
-            if debug:
-                table_debug_pages.append(
-                    _collect_table_drawing_debug(
-                        page,
-                        page_no=page_idx,
-                        header_margin=header_margin,
-                        footer_margin=footer_margin,
-                    )
-                )
-                edge_debug_pages.append(
-                    _collect_page_edge_debug(
-                        page,
-                        page_no=page_idx,
-                        header_margin=header_margin,
-                        footer_margin=footer_margin,
-                    )
-                )
-            if debug_watermark:
-                rotated_debug.extend(_collect_rotated_text_debug(page, page_no=page_idx))
-            tables = _extract_tables(page, force_table=force_table)
-            full_page_text = _extract_body_text(
-                page,
-                header_margin=header_margin,
-                footer_margin=footer_margin,
-            )
-            page_text = _extract_body_text(
-                page,
-                header_margin=header_margin,
-                footer_margin=footer_margin,
-                excluded_bboxes=_body_excluded_bboxes(pending_table, tables, footer_margin),
-            )
-            if page_text.strip():
-                output_text.append(f"### Page {page_idx}\n{page_text}")
-
-            if tables:
-                body_top, body_bottom = _detect_body_bounds(
-                    page,
-                    header_margin=header_margin,
-                    footer_margin=footer_margin,
-                )
-                for table_rows, bbox in tables:
-                    table_bboxes = [table_bbox for _table_rows, table_bbox in tables]
-                    cross_page_continuation = _should_try_table_continuation_merge(
-                        pending_page=pending_last_page,
-                        current_page=page_idx,
-                    )
-
-                    merged_missing_first = None
-                    if cross_page_continuation:
-                        merged_missing_first = _maybe_merge_missing_first_column_chunk(
-                            pending_table,
-                            table_rows,
-                            full_page_text,
-                        )
-                    if merged_missing_first is not None:
-                        pending_table = merged_missing_first
-                        pending_last_page = page_idx
-                        pending_bbox = bbox
-                        pending_axes = _vertical_axes_for_bbox(page, bbox)
-                        pending_gap_text_boxes = _gap_text_boxes_after_bbox(
-                            page,
-                            bbox,
-                            table_bboxes,
-                            header_margin=header_margin,
-                            footer_margin=footer_margin,
-                        )
-                        continue
-
-                    continuation_rows = table_rows
-                    if cross_page_continuation:
-                        continuation_rows = _split_repeated_header(pending_table or [], table_rows)
-                        if pending_table is not None and _is_continuation_chunk(pending_table, continuation_rows):
-                            pending_table.extend(continuation_rows)
-                            pending_last_page = page_idx
-                            current_axes = _vertical_axes_for_bbox(page, bbox)
-                            if pending_bbox is not None:
-                                pending_bbox = (
-                                    min(pending_bbox[0], bbox[0]),
-                                    min(pending_bbox[1], bbox[1]),
-                                    max(pending_bbox[2], bbox[2]),
-                                    max(pending_bbox[3], bbox[3]),
-                                )
-                            else:
-                                    pending_bbox = bbox
-                            pending_axes = _merge_numeric_positions([*pending_axes, *current_axes], tolerance=1.0)
-                            pending_gap_text_boxes = _gap_text_boxes_after_bbox(
-                                page,
-                                bbox,
-                                table_bboxes,
-                                header_margin=header_margin,
-                                footer_margin=footer_margin,
-                            )
-                            continue
-
-                    current_axes = _vertical_axes_for_bbox(page, bbox)
-                    current_gap_text_boxes = _gap_text_boxes_before_bbox(
-                        page,
-                        bbox,
-                        table_bboxes,
-                        header_margin=header_margin,
-                        footer_margin=footer_margin,
-                    )
-                    if (
-                        pending_table is not None
-                        and pending_bbox is not None
-                        and pending_last_page is not None
-                        and cross_page_continuation
-                        and _continuation_regions_should_merge(
-                            prev_bbox=pending_bbox,
-                            curr_bbox=bbox,
-                            prev_axes=pending_axes,
-                            curr_axes=current_axes,
-                            body_top=body_top,
-                            body_bottom=body_bottom,
-                            gap_text_boxes=[*pending_gap_text_boxes, *current_gap_text_boxes],
-                        )
-                    ):
-                        pending_table.extend(continuation_rows)
-                        pending_last_page = page_idx
-                        pending_bbox = (
-                            min(pending_bbox[0], bbox[0]),
-                            min(pending_bbox[1], bbox[1]),
-                            max(pending_bbox[2], bbox[2]),
-                            max(pending_bbox[3], bbox[3]),
-                        )
-                        pending_axes = _merge_numeric_positions([*pending_axes, *current_axes], tolerance=1.0)
-                        pending_gap_text_boxes = _gap_text_boxes_after_bbox(
-                            page,
-                            bbox,
-                            table_bboxes,
-                            header_margin=header_margin,
-                            footer_margin=footer_margin,
-                        )
-                        continue
-
-                    _flush_pending()
-                    pending_table = table_rows
-                    pending_page = page_idx
-                    pending_last_page = page_idx
-                    pending_bbox = bbox
-                    pending_axes = current_axes
-                    pending_gap_text_boxes = _gap_text_boxes_after_bbox(
-                        page,
-                        bbox,
-                        table_bboxes,
-                        header_margin=header_margin,
-                        footer_margin=footer_margin,
-                    )
-
-        _flush_pending()
-
-    markdown = "\n\n".join(output_text)
-    table_markdown = "\n\n".join(output_tables)
-
-    text_file = out_md_dir / f"{stem}.txt"
-    md_file = out_md_dir / f"{stem}.md"
-    table_md_file = out_md_dir / f"{stem}_table.md"
-    text_file.write_text(markdown, encoding="utf-8")
-    md_file.write_text(markdown, encoding="utf-8")
-    table_md_file.write_text(table_markdown, encoding="utf-8")
-
-    image_files = _extract_embedded_images(
-        pdf_path=pdf_path,
-        out_image_dir=out_image_dir,
-        stem=stem,
-        pages=pages,
-    )
-
-    summary = {
-        "pdf": str(pdf_path),
-        "text_file": str(text_file),
-        "md_file": str(md_file),
-        "table_md_file": str(table_md_file),
-        "images": [str(p) for p in image_files],
-        "table_count": len(output_tables),
-    }
-    summary_file = out_md_dir / f"{stem}_summary.json"
-    summary_file.write_text(json.dumps(summary, ensure_ascii=False, indent=2), encoding="utf-8")
-
-    debug_file: Optional[Path] = None
-    debug_edges_file: Optional[Path] = None
-    if debug:
-        debug_file = out_md_dir / f"{stem}_debug.json"
-        debug_payload = {
-            "pdf": str(pdf_path),
-            "pages": table_debug_pages,
-        }
-        debug_file.write_text(json.dumps(debug_payload, ensure_ascii=False, indent=2), encoding="utf-8")
-        debug_edges_file = out_md_dir / f"{stem}_edges_debug.json"
-        debug_edges_payload = {
-            "pdf": str(pdf_path),
-            "pages": edge_debug_pages,
-        }
-        debug_edges_file.write_text(json.dumps(debug_edges_payload, ensure_ascii=False, indent=2), encoding="utf-8")
-
-    debug_watermark_file: Optional[Path] = None
-    if debug_watermark:
-        debug_watermark_file = out_md_dir / f"{stem}_watermark_debug.json"
-        debug_watermark_file.write_text(json.dumps(rotated_debug, ensure_ascii=False, indent=2), encoding="utf-8")
-
-    return {
-        "markdown": markdown,
-        "table_markdown": table_markdown,
-        "text_file": text_file,
-        "md_file": md_file,
-        "table_md_file": table_md_file,
-        "debug_file": debug_file,
-        "debug_edges_file": debug_edges_file,
-        "debug_watermark_file": debug_watermark_file,
-        "image_files": image_files,
-        "summary": summary,
-    }
-
-
-if __name__ == "__main__":  # basic manual run
-    import argparse
-
-    parser = argparse.ArgumentParser()
-    parser.add_argument("pdf_path")
-    parser.add_argument("--out-md-dir", default="graph_pdf/artifacts/md")
-    parser.add_argument("--out-image-dir", default="graph_pdf/artifacts/images")
-    parser.add_argument("--stem", default="output")
-    parser.add_argument("--pages", help="1-based pages like 1,3,5-8")
-    parser.add_argument("--force-table", action="store_true")
-    parser.add_argument("--debug", action="store_true")
-    parser.add_argument("--debug-watermark", action="store_true")
-    args = parser.parse_args()
-
-    extract_pdf_to_outputs(
-        pdf_path=Path(args.pdf_path),
-        out_md_dir=Path(args.out_md_dir),
-        out_image_dir=Path(args.out_image_dir),
-        stem=args.stem,
-        pages=_parse_pages_spec(args.pages) if args.pages else None,
-        force_table=args.force_table,
-        debug=args.debug,
-        debug_watermark=args.debug_watermark,
-    )
